+++ src/serial.rs
@@ -95,7 +95,7 @@ usart_pins! {
         rx => [gpioa::PA3<Alternate<AF1>>, gpioa::PA15<Alternate<AF1>>],
     }
 }
-#[cfg(feature = "stm32f070")]
+#[cfg(any(feature = "stm32f070", feature = "stm32f072"))]
 usart_pins! {
     USART1 => {
         tx => [gpioa::PA9<Alternate<AF1>>, gpiob::PB6<Alternate<AF0>>],
@@ -107,6 +107,7 @@ usart_pins! {
     feature = "stm32f030xc",
     feature = "stm32f042",
     feature = "stm32f070",
+    feature = "stm32f072",
 ))]
 usart_pins! {
     USART2 => {
@@ -114,7 +115,18 @@ usart_pins! {
         rx => [gpioa::PA3<Alternate<AF1>>, gpioa::PA15<Alternate<AF1>>],
     }
 }
-#[cfg(any(feature = "stm32f030xc", feature = "stm32f070xb"))]
+#[cfg(feature = "stm32f072")]
+usart_pins! {
+    USART2 => {
+        tx => [gpiod::PD5<Alternate<AF0>>],
+        rx => [gpiod::PD6<Alternate<AF0>>],
+    }
+}
+#[cfg(any(
+    feature = "stm32f030xc",
+    feature = "stm32f070xb",
+    feature = "stm32f072"
+))]
 usart_pins! {
     USART3 => {
         // According to the datasheet PB10 is both tx and rx, but in stm32cubemx it's only tx
@@ -124,6 +136,13 @@ usart_pins! {
     USART4 => {
         tx => [gpioa::PA0<Alternate<AF4>>, gpioc::PC10<Alternate<AF0>>],
         rx => [gpioa::PA1<Alternate<AF4>>, gpioc::PC11<Alternate<AF0>>],
+    }
+}
+#[cfg(feature = "stm32f072")]
+usart_pins! {
+    USART3 => {
+        tx => [gpiod::PD8<Alternate<AF0>>],
+        rx => [gpiod::PD9<Alternate<AF0>>],
     }
 }
 #[cfg(feature = "stm32f030xc")]
@@ -240,11 +259,16 @@ usart! {
     feature = "stm32f030xc",
     feature = "stm32f042",
     feature = "stm32f070",
+    feature = "stm32f072",
 ))]
 usart! {
     USART2: (usart2, usart2en, apb1enr),
 }
-#[cfg(any(feature = "stm32f030xc", feature = "stm32f070xb"))]
+#[cfg(any(
+    feature = "stm32f030xc",
+    feature = "stm32f070xb",
+    feature = "stm32f072",
+))]
 usart! {
     USART3: (usart3, usart3en, apb1enr),
     USART4: (usart4, usart4en, apb1enr),