+++ src/spi.rs
@@ -13,7 +13,8 @@ use crate::stm32::SPI1;
 #[cfg(any(
     feature = "stm32f030x8",
     feature = "stm32f030xc",
-    feature = "stm32f070xb"
+    feature = "stm32f070xb",
+    feature = "stm32f072",
 ))]
 use crate::stm32::SPI2;
 
@@ -87,10 +88,23 @@ spi_pins! {
         mosi => [gpiob::PB15<Alternate<AF0>>],
     }
 }
+
+// TODO: The ST SVD files are missing the entire PE enable register.
+//       So those pins do not exist in the register definitions.
+//       Re-enable as soon as this gets fixed.
+// #[cfg(feature = "stm32f072")]
+// spi_pins! {
+//     SPI1 => {
+//         sck => [gpioe::PE13<Alternate<AF1>>],
+//         miso => [gpioe::PE14<Alternate<AF1>>],
+//         mosi => [gpioe::PE15<Alternate<AF1>>],
+//     }
+// }
 #[cfg(any(
     feature = "stm32f030x8",
     feature = "stm32f030xc",
-    feature = "stm32f070xb"
+    feature = "stm32f070xb",
+    feature = "stm32f072",
 ))]
 spi_pins! {
     SPI2 => {
@@ -99,12 +113,24 @@ spi_pins! {
         mosi => [gpiob::PB15<Alternate<AF0>>],
     }
 }
-#[cfg(any(feature = "stm32f030xc", feature = "stm32f070xb"))]
+#[cfg(any(
+    feature = "stm32f030xc",
+    feature = "stm32f070xb",
+    feature = "stm32f072",
+))]
 spi_pins! {
     SPI2 => {
         sck => [gpiob::PB10<Alternate<AF5>>],
         miso => [gpioc::PC2<Alternate<AF1>>],
         mosi => [gpioc::PC3<Alternate<AF1>>],
+    }
+}
+#[cfg(feature = "stm32f072")]
+spi_pins! {
+    SPI2 => {
+        sck => [gpiod::PD1<Alternate<AF1>>],
+        miso => [gpiod::PD3<Alternate<AF1>>],
+        mosi => [gpiod::PD4<Alternate<AF1>>],
     }
 }