+++ CHANGELOG.md
@@ -9,6 +9,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 
+- Added peripheral definitions for the stm32f072xx line - @Yatekii
+
+### Changed
+
+- Fixed broken PC GPIO definitions with feature = "stm32f030" and feature = "stm32f070"
 - More robust error handling for I2C
 
 ## [v0.11.0] - 2019-01-04